Add some fixes so you can't change games you don't own.

+++ app.py
@@ -1,3 +1,4 @@
+from os import abort
 import sqlite3
 from flask import Flask
 from flask import redirect, render_template, request, session
@@ -30,6 +31,8 @@ def new_game():
 @app.route("/edit_game/<int:game_id>")
 def edit_game(game_id):
     game = games.get_game(game_id)
+    if game["user_id"] != session.get("user_id"):
+        abort(403)
     return render_template("edit_game.html", game=game)
 
 
@@ -46,6 +49,9 @@ def find_game():
 
 @app.route("/delete_game/<int:game_id>", methods=["GET", "POST"])
 def delete_game(game_id):
+    game = games.get_game(game_id)
+    if game["user_id"] != session.get("user_id"):
+        abort(403)
     if request.method == "POST":
         if "confirm" in request.form:
             games.delete_game(game_id)
@@ -59,7 +65,11 @@ def delete_game(game_id):
 
 @app.route("/update_game", methods=["POST"])
 def update_game():
+
     game_id = request.form["game_id"]
+    game = games.get_game(game_id)
+    if game["user_id"] != session.get("user_id"):
+        abort(403)
     title = request.form["title"]
     description = request.form["description"]
     date = request.form["date"]
